+++ test/fixtures/plugins/plugin-a/package.json
