+++ test/fixtures/plugins/plugin-b/package.json
