+++ test/fixtures/plugins/plugin-c/package.json
