+++ test/fixtures/plugins/plugin-wrong-a/package.json
