+++ test/fixtures/plugins/plugin-wrong-b/package.json
