+++ test/fixtures/trigger/event/tsconfig.json
