+++ test/fixtures/trigger/http/httpTrigger.ts
@@ -1,7 +1,7 @@
 import { Stream } from 'stream';
 import { Context, Next } from '@artus/pipeline';
-import { Trigger } from '../../../src';
-import { DefineTrigger } from '../../../src/decorator';
+import { Trigger } from '../../../../src';
+import { DefineTrigger } from '../../../../src/decorator';
 
 @DefineTrigger()
 export default class HttpTrigger extends Trigger {
