+++ test/fixtures/trigger/timer/tsconfig.json
