+++ test/trigger/http.test.ts
@@ -7,7 +7,7 @@ describe('test/trigger/http.test.ts', () => {
     const {
       main,
       isListening
-    } = await import('../fixtures/trigger-http/app');
+    } = await import('../fixtures/trigger/http/app');
     const app = await main();
     const testResponse = await axios.get('http://127.0.0.1:3001');
     await axios.get('http://127.0.0.1:3001');
